--- stream-core/src/main/java/io/github/vampireachao/stream/core/stream/FastStream.java
@@ -1,1569 +1,0 @@
-package io.github.vampireachao.stream.core.stream;
-
-
-import io.github.vampireachao.stream.core.collector.Collectors;
-import io.github.vampireachao.stream.core.optional.Opp;
-
-import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.*;
-import java.util.stream.*;
-
-/**
- * 对Stream的封装和拓展
- *
- * @author VampireAchao
- */
-public class FastStream<T> implements Stream<T>, Iterable<T> {
-
-    protected final Stream<T> stream;
-
-    FastStream(Stream<T> stream) {
-        this.stream = stream;
-    }
-
-    /**
-     * Returns a builder for a {@code Stream}.
-     *
-     * @param <T> type of elements
-     * @return a stream builder
-     */
-    public static <T> StreamBuilder<T> builder() {
-        return new StreamBuilder<T>() {
-            private final Builder<T> streamBuilder = Stream.builder();
-
-            @Override
-            public void accept(T t) {
-                streamBuilder.accept(t);
-            }
-
-            @Override
-            public FastStream<T> build() {
-                return new FastStream<>(streamBuilder.build());
-            }
-        };
-    }
-
-    /**
-     * Returns an empty sequential {@code Stream}.
-     *
-     * @param <T> the type of stream elements
-     * @return an empty sequential stream
-     */
-    public static <T> FastStream<T> empty() {
-        return new FastStream<>(Stream.empty());
-    }
-
-    /**
-     * Returns a sequential {@code Stream} containing a single element.
-     *
-     * @param t   the single element
-     * @param <T> the type of stream elements
-     * @return a singleton sequential stream
-     */
-    public static <T> FastStream<T> of(T t) {
-        return new FastStream<>(Stream.of(t));
-    }
-
-    /**
-     * Returns a sequential ordered stream whose elements are the specified values.
-     *
-     * @param <T>    the type of stream elements
-     * @param values the elements of the new stream
-     * @return the new stream
-     * Creating a stream from an array is safe
-     */
-    @SafeVarargs
-    @SuppressWarnings("varargs")
-    public static <T> FastStream<T> of(T... values) {
-        return new FastStream<>(Arrays.stream(values));
-    }
-
-    /**
-     * Returns an infinite sequential ordered {@code Stream} produced by iterative
-     * application of a function {@code f} to an initial element {@code seed},
-     * producing a {@code Stream} consisting of {@code seed}, {@code f(seed)},
-     * {@code f(f(seed))}, etc.
-     *
-     * <p>The first element (position {@code 0}) in the {@code Stream} will be
-     * the provided {@code seed}.  For {@code n > 0}, the element at position
-     * {@code n}, will be the result of applying the function {@code f} to the
-     * element at position {@code n - 1}.
-     *
-     * @param <T>  the type of stream elements
-     * @param seed the initial element
-     * @param f    a function to be applied to the previous element to produce
-     *             a new element
-     * @return a new sequential {@code Stream}
-     */
-    public static <T> FastStream<T> iterate(final T seed, final UnaryOperator<T> f) {
-        return new FastStream<>(Stream.iterate(seed, f));
-    }
-
-    public static <T> FastStream<T> iterate(T seed, Predicate<? super T> hasNext, UnaryOperator<T> next) {
-        Objects.requireNonNull(next);
-        Objects.requireNonNull(hasNext);
-        Spliterator<T> spliterator = new Spliterators.AbstractSpliterator<T>(Long.MAX_VALUE,
-                Spliterator.ORDERED | Spliterator.IMMUTABLE) {
-            T prev;
-            boolean started;
-            boolean finished;
-
-            @Override
-            public boolean tryAdvance(Consumer<? super T> action) {
-                Objects.requireNonNull(action);
-                if (finished) {
-                    return false;
-                }
-                T t;
-                if (started) {
-                    t = next.apply(prev);
-                } else {
-                    t = seed;
-                    started = true;
-                }
-                if (!hasNext.test(t)) {
-                    prev = null;
-                    finished = true;
-                    return false;
-                }
-                prev = t;
-                action.accept(prev);
-                return true;
-            }
-
-            @Override
-            public void forEachRemaining(Consumer<? super T> action) {
-                Objects.requireNonNull(action);
-                if (finished) {
-                    return;
-                }
-                finished = true;
-                T t = started ? next.apply(prev) : seed;
-                prev = null;
-                while (hasNext.test(t)) {
-                    action.accept(t);
-                    t = next.apply(t);
-                }
-            }
-        };
-        return new FastStream<>(StreamSupport.stream(spliterator, false));
-    }
-
-    /**
-     * Returns an infinite sequential unordered stream where each element is
-     * generated by the provided {@code Supplier}.  This is suitable for
-     * generating constant streams, streams of random elements, etc.
-     *
-     * @param <T> the type of stream elements
-     * @param s   the {@code Supplier} of generated elements
-     * @return a new infinite sequential unordered {@code Stream}
-     */
-    public static <T> FastStream<T> generate(Supplier<T> s) {
-        return new FastStream<>(Stream.generate(s));
-    }
-
-    /**
-     * Creates a lazily concatenated stream whose elements are all the
-     * elements of the first stream followed by all the elements of the
-     * second stream.  The resulting stream is ordered if both
-     * of the input streams are ordered, and parallel if either of the input
-     * streams is parallel.  When the resulting stream is closed, the close
-     * handlers for both input streams are invoked.
-     *
-     * @param <T> The type of stream elements
-     * @param a   the first stream
-     * @param b   the second stream
-     * @return the concatenation of the two input streams
-     * @implNote Use caution when constructing streams from repeated concatenation.
-     * Accessing an element of a deeply concatenated stream can result in deep
-     * call chains, or even {@code StackOverflowException}.
-     */
-    public static <T> FastStream<T> concat(FastStream<? extends T> a, FastStream<? extends T> b) {
-        return new FastStream<>(Stream.concat(a, b));
-    }
-
-    public static <T> FastStream<T> of(Iterable<T> iterable) {
-        return of(iterable, false);
-    }
-
-    public static <T> FastStream<T> of(Iterable<T> iterable, boolean parallel) {
-        return Optional.ofNullable(iterable).map(Iterable::spliterator).map(spliterator -> StreamSupport.stream(spliterator, parallel)).map(FastStream::new).orElseGet(FastStream::empty);
-    }
-
-    public static <T> FastStream<T> of(Stream<T> stream) {
-        return new FastStream<>(stream);
-    }
-
-    public static FastStream<String> split(CharSequence str, String regex) {
-        return Opp.ofBlankAble(str).map(String::valueOf).map(s -> s.split(regex)).map(FastStream::of).orElseGet(FastStream::empty);
-    }
-
-    /**
-     * Returns a stream consisting of the elements of this stream that match
-     * the given predicate.
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param predicate a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                  <a href="package-summary.html#Statelessness">stateless</a>
-     *                  predicate to apply to each element to determine if it
-     *                  should be included
-     * @return the new stream
-     */
-    @Override
-    public FastStream<T> filter(Predicate<? super T> predicate) {
-        return new FastStream<>(stream.filter(predicate));
-    }
-
-    public <R> FastStream<T> filter(Function<? super T, ? extends R> mapper, R value) {
-        return filter(e -> Objects.equals(Opp.ofNullable(e).map(mapper).get(), value));
-    }
-
-    public FastStream<T> filterIdx(BiPredicate<? super T, Integer> predicate) {
-        AtomicInteger index = new AtomicInteger(-1);
-        return filter(e -> predicate.test(e, isParallel() ? index.get() : index.incrementAndGet()));
-    }
-
-    public FastStream<T> nonNull() {
-        return new FastStream<>(stream.filter(Objects::nonNull));
-    }
-
-    /**
-     * Returns a stream consisting of the results of applying the given
-     * function to the elements of this stream.
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element
-     * @return the new stream
-     */
-    @Override
-    public <R> FastStream<R> map(Function<? super T, ? extends R> mapper) {
-        return new FastStream<>(stream.map(mapper));
-    }
-
-    public <R> FastStream<R> mapIdx(BiFunction<? super T, Integer, ? extends R> mapper) {
-        AtomicInteger index = new AtomicInteger(-1);
-        return map(e -> mapper.apply(e, isParallel() ? index.get() : index.incrementAndGet()));
-    }
-
-    /**
-     * Returns a stream consisting of the results of replacing each element of
-     * this stream with the contents of a mapped stream produced by applying
-     * the provided mapping function to each element.  Each mapped stream is
-     * {@link FastStream#close() closed} after its contents
-     * have been placed into this stream.  (If a mapped stream is {@code null}
-     * an empty stream is used, instead.)
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element which produces a stream
-     *               of new values
-     * @return the new stream
-     * @apiNote The {@code flatMap()} operation has the effect of applying a one-to-many
-     * transformation to the elements of the stream, and then flattening the
-     * resulting elements into a new stream.
-     *
-     * <p><b>Examples.</b>
-     *
-     * <p>If {@code orders} is a stream of purchase orders, and each purchase
-     * order contains a collection of line items, then the following produces a
-     * stream containing all the line items in all the orders:
-     * <pre>{@code
-     *     orders.flatMap(order -> order.getLineItems().stream())...
-     * }</pre>
-     *
-     * <p>If {@code path} is the path to a file, then the following produces a
-     * stream of the {@code words} contained in that file:
-     * <pre>{@code
-     *     Steam<String> lines = Files.lines(path, StandardCharsets.UTF_8);
-     *     Steam<String> words = lines.flatMap(line -> Steam.of(line.split(" +")));
-     * }</pre>
-     * The {@code mapper} function passed to {@code flatMap} splits a line,
-     * using a simple regular expression, into an array of words, and then
-     * creates a stream of words from that array.
-     */
-    @Override
-    public <R> FastStream<R> flatMap(Function<? super T, ? extends Stream<? extends R>> mapper) {
-        return new FastStream<>(stream.flatMap(mapper));
-    }
-
-    public <R> FastStream<R> flatMapIdx(BiFunction<? super T, Integer, ? extends Stream<? extends R>> mapper) {
-        AtomicInteger index = new AtomicInteger(-1);
-        return flatMap(e -> mapper.apply(e, isParallel() ? index.get() : index.incrementAndGet()));
-    }
-
-    /**
-     * Returns an {@code IntBaseSteam} consisting of the results of applying the
-     * given function to the elements of this stream.
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">
-     * intermediate operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element
-     * @return the new stream
-     */
-    @Override
-    public IntStream mapToInt(ToIntFunction<? super T> mapper) {
-        return stream.mapToInt(mapper);
-    }
-
-    /**
-     * Returns a {@code LongBaseSteam} consisting of the results of applying the
-     * given function to the elements of this stream.
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element
-     * @return the new stream
-     */
-    @Override
-    public LongStream mapToLong(ToLongFunction<? super T> mapper) {
-        return stream.mapToLong(mapper);
-    }
-
-    /**
-     * Returns a {@code DoubleBaseSteam} consisting of the results of applying the
-     * given function to the elements of this stream.
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element
-     * @return the new stream
-     */
-    @Override
-    public DoubleStream mapToDouble(ToDoubleFunction<? super T> mapper) {
-        return stream.mapToDouble(mapper);
-    }
-
-    public <R> FastStream<R> flatMapIter(Function<? super T, ? extends Iterable<? extends R>> mapper) {
-        return flatMap(w -> Opp.of(w).map(mapper).map(FastStream::of).orElseGet(FastStream::empty));
-    }
-
-    /**
-     * Returns an {@code IntBaseSteam} consisting of the results of replacing each
-     * element of this stream with the contents of a mapped stream produced by
-     * applying the provided mapping function to each element.  Each mapped
-     * stream is {@link FastStream#close() closed} after its
-     * contents have been placed into this stream.  (If a mapped stream is
-     * {@code null} an empty stream is used, instead.)
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element which produces a stream
-     *               of new values
-     * @return the new stream
-     * @see #flatMap(Function)
-     */
-    @Override
-    public IntStream flatMapToInt(Function<? super T, ? extends IntStream> mapper) {
-        return stream.flatMapToInt(mapper);
-    }
-
-    /**
-     * Returns an {@code LongBaseSteam} consisting of the results of replacing each
-     * element of this stream with the contents of a mapped stream produced by
-     * applying the provided mapping function to each element.  Each mapped
-     * stream is {@link FastStream#close() closed} after its
-     * contents have been placed into this stream.  (If a mapped stream is
-     * {@code null} an empty stream is used, instead.)
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element which produces a stream
-     *               of new values
-     * @return the new stream
-     * @see #flatMap(Function)
-     */
-    @Override
-    public LongStream flatMapToLong(Function<? super T, ? extends LongStream> mapper) {
-        return stream.flatMapToLong(mapper);
-    }
-
-    /**
-     * Returns an {@code DoubleBaseSteam} consisting of the results of replacing
-     * each element of this stream with the contents of a mapped stream produced
-     * by applying the provided mapping function to each element.  Each mapped
-     * stream is {@link FastStream#close() closed} after its
-     * contents have placed been into this stream.  (If a mapped stream is
-     * {@code null} an empty stream is used, instead.)
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * @param mapper a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *               <a href="package-summary.html#Statelessness">stateless</a>
-     *               function to apply to each element which produces a stream
-     *               of new values
-     * @return the new stream
-     * @see #flatMap(Function)
-     */
-    @Override
-    public DoubleStream flatMapToDouble(Function<? super T, ? extends DoubleStream> mapper) {
-        return stream.flatMapToDouble(mapper);
-    }
-
-    public <R> FastStream<R> mapMulti(BiConsumer<? super T, ? super Consumer<R>> mapper) {
-        Objects.requireNonNull(mapper);
-        return flatMap(e -> {
-            StreamBuilder<R> buffer = FastStream.builder();
-            mapper.accept(e, buffer);
-            return buffer.build();
-        });
-    }
-
-    /**
-     * Returns a stream consisting of the distinct elements (according to
-     * {@link Object#equals(Object)}) of this stream.
-     *
-     * <p>For ordered streams, the selection of distinct elements is stable
-     * (for duplicated elements, the element appearing first in the encounter
-     * order is preserved.)  For unordered streams, no stability guarantees
-     * are made.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">stateful
-     * intermediate operation</a>.
-     *
-     * @return the new stream
-     * @apiNote Preserving stability for {@code distinct()} in parallel pipelines is
-     * relatively expensive (requires that the operation act as a full barrier,
-     * with substantial buffering overhead), and stability is often not needed.
-     * Using an unordered stream source (such as {@link #generate(Supplier)})
-     * or removing the ordering constraint with {@link #unordered()} may result
-     * in significantly more efficient execution for {@code distinct()} in parallel
-     * pipelines, if the semantics of your situation permit.  If consistency
-     * with encounter order is required, and you are experiencing poor performance
-     * or memory utilization with {@code distinct()} in parallel pipelines,
-     * switching to sequential execution with {@link #sequential()} may improve
-     * performance.
-     */
-    @Override
-    public FastStream<T> distinct() {
-        return new FastStream<>(stream.distinct());
-    }
-
-    public FastStream<T> distinct(Function<? super T, ?> keyExtractor) {
-        return new FastStream<>(toMap(keyExtractor).entrySet().stream()).parallel(isParallel()).map(Map.Entry::getValue);
-    }
-
-    /**
-     * Returns a stream consisting of the elements of this stream, sorted
-     * according to natural order.  If the elements of this stream are not
-     * {@code Comparable}, a {@code java.lang.ClassCastException} may be thrown
-     * when the terminal operation is executed.
-     *
-     * <p>For ordered streams, the sort is stable.  For unordered streams, no
-     * stability guarantees are made.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">stateful
-     * intermediate operation</a>.
-     *
-     * @return the new stream
-     */
-    @Override
-    public FastStream<T> sorted() {
-        return new FastStream<>(stream.sorted());
-    }
-
-    /**
-     * Returns a stream consisting of the elements of this stream, sorted
-     * according to the provided {@code Comparator}.
-     *
-     * <p>For ordered streams, the sort is stable.  For unordered streams, no
-     * stability guarantees are made.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">stateful
-     * intermediate operation</a>.
-     *
-     * @param comparator a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                   <a href="package-summary.html#Statelessness">stateless</a>
-     *                   {@code Comparator} to be used to compare stream elements
-     * @return the new stream
-     */
-    @Override
-    public FastStream<T> sorted(Comparator<? super T> comparator) {
-        return new FastStream<>(stream.sorted(comparator));
-    }
-
-    /**
-     * Returns a stream consisting of the elements of this stream, additionally
-     * performing the provided action on each element as elements are consumed
-     * from the resulting stream.
-     *
-     * <p>This is an <a href="package-summary.html#BaseSteamOps">intermediate
-     * operation</a>.
-     *
-     * <p>For parallel stream pipelines, the action may be called at
-     * whatever time and in whatever thread the element is made available by the
-     * upstream operation.  If the action modifies shared state,
-     * it is responsible for providing the required synchronization.
-     *
-     * @param action a <a href="package-summary.html#NonInterference">
-     *               non-interfering</a> action to perform on the elements as
-     *               they are consumed from the stream
-     * @return the new stream
-     * @apiNote This method exists mainly to support debugging, where you want
-     * to see the elements as they flow past a certain point in a pipeline:
-     * <pre>{@code
-     *     Steam.of("one", "two", "three", "four")
-     *         .filter(e -> e.length() > 3)
-     *         .peek(e -> System.out.println("Filtered value: " + e))
-     *         .map(String::toUpperCase)
-     *         .peek(e -> System.out.println("Mapped value: " + e))
-     *         .collect(Collectors.toList());
-     * }</pre>
-     */
-    @Override
-    public FastStream<T> peek(Consumer<? super T> action) {
-        return new FastStream<>(stream.peek(action));
-    }
-
-    /**
-     * Returns a stream consisting of the elements of this stream, truncated
-     * to be no longer than {@code maxSize} in length.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">short-circuiting
-     * stateful intermediate operation</a>.
-     *
-     * @param maxSize the number of elements the stream should be limited to
-     * @return the new stream
-     * @throws IllegalArgumentException if {@code maxSize} is negative
-     * @apiNote While {@code limit()} is generally a cheap operation on sequential
-     * stream pipelines, it can be quite expensive on ordered parallel pipelines,
-     * especially for large values of {@code maxSize}, since {@code limit(n)}
-     * is constrained to return not just any <em>n</em> elements, but the
-     * <em>first n</em> elements in the encounter order.  Using an unordered
-     * stream source (such as {@link #generate(Supplier)}) or removing the
-     * ordering constraint with {@link #unordered()} may result in significant
-     * speedups of {@code limit()} in parallel pipelines, if the semantics of
-     * your situation permit.  If consistency with encounter order is required,
-     * and you are experiencing poor performance or memory utilization with
-     * {@code limit()} in parallel pipelines, switching to sequential execution
-     * with {@link #sequential()} may improve performance.
-     */
-    @Override
-    public FastStream<T> limit(long maxSize) {
-        return new FastStream<>(stream.limit(maxSize));
-    }
-
-    /**
-     * Returns a stream consisting of the remaining elements of this stream
-     * after discarding the first {@code n} elements of the stream.
-     * If this stream contains fewer than {@code n} elements then an
-     * empty stream will be returned.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">stateful
-     * intermediate operation</a>.
-     *
-     * @param n the number of leading elements to skip
-     * @return the new stream
-     * @throws IllegalArgumentException if {@code n} is negative
-     * @apiNote While {@code skip()} is generally a cheap operation on sequential
-     * stream pipelines, it can be quite expensive on ordered parallel pipelines,
-     * especially for large values of {@code n}, since {@code skip(n)}
-     * is constrained to skip not just any <em>n</em> elements, but the
-     * <em>first n</em> elements in the encounter order.  Using an unordered
-     * stream source (such as {@link #generate(Supplier)}) or removing the
-     * ordering constraint with {@link #unordered()} may result in significant
-     * speedups of {@code skip()} in parallel pipelines, if the semantics of
-     * your situation permit.  If consistency with encounter order is required,
-     * and you are experiencing poor performance or memory utilization with
-     * {@code skip()} in parallel pipelines, switching to sequential execution
-     * with {@link #sequential()} may improve performance.
-     */
-    @Override
-    public FastStream<T> skip(long n) {
-        return new FastStream<>(stream.skip(n));
-    }
-
-    /**
-     * Returns an equivalent stream that is sequential.  May return
-     * itself, either because the stream was already sequential, or because
-     * the underlying stream state was modified to be sequential.
-     *
-     * <p>This is an <a href="package-summary.html#StreamOps">intermediate
-     * operation</a>.
-     *
-     * @return a sequential stream
-     */
-    @Override
-    public FastStream<T> sequential() {
-        return new FastStream<>(stream.sequential());
-    }
-
-    /**
-     * Performs an action for each element of this stream.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * <p>The behavior of this operation is explicitly nondeterministic.
-     * For parallel stream pipelines, this operation does <em>not</em>
-     * guarantee to respect the encounter order of the stream, as doing so
-     * would sacrifice the benefit of parallelism.  For any given element, the
-     * action may be performed at whatever time and in whatever thread the
-     * library chooses.  If the action accesses shared state, it is
-     * responsible for providing the required synchronization.
-     *
-     * @param action a <a href="package-summary.html#NonInterference">
-     *               non-interfering</a> action to perform on the elements
-     */
-    @Override
-    public void forEach(Consumer<? super T> action) {
-        stream.forEach(action);
-    }
-
-    public void forEachIdx(BiConsumer<? super T, Integer> action) {
-        AtomicInteger index = new AtomicInteger(-1);
-        stream.forEach(e -> action.accept(e, isParallel() ? index.get() : index.incrementAndGet()));
-    }
-
-    /**
-     * Performs an action for each element of this stream, in the encounter
-     * order of the stream if the stream has a defined encounter order.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * <p>This operation processes the elements one at a time, in encounter
-     * order if one exists.  Performing the action for one element
-     * <a href="../concurrent/package-summary.html#MemoryVisibility"><i>happens-before</i></a>
-     * performing the action for subsequent elements, but for any given element,
-     * the action may be performed in whatever thread the library chooses.
-     *
-     * @param action a <a href="package-summary.html#NonInterference">
-     *               non-interfering</a> action to perform on the elements
-     * @see #forEach(Consumer)
-     */
-    @Override
-    public void forEachOrdered(Consumer<? super T> action) {
-        stream.forEachOrdered(action);
-    }
-
-    public void forEachOrderedIdx(BiConsumer<? super T, Integer> action) {
-        AtomicInteger index = new AtomicInteger(-1);
-        stream.forEachOrdered(e -> action.accept(e, isParallel() ? index.get() : index.incrementAndGet()));
-    }
-
-    /**
-     * Returns an array containing the elements of this stream.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @return an array containing the elements of this stream
-     */
-    @Override
-    public Object[] toArray() {
-        return stream.toArray();
-    }
-
-    /**
-     * Returns an array containing the elements of this stream, using the
-     * provided {@code generator} function to allocate the returned array, as
-     * well as any additional arrays that might be required for a partitioned
-     * execution or for resizing.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @param generator a function which produces a new array of the desired
-     *                  type and the provided length
-     * @return an array containing the elements in this stream
-     * @throws ArrayStoreException if the runtime type of the array returned
-     *                             from the array generator is not a supertype of the runtime type of every
-     *                             element in this stream
-     * @apiNote The generator function takes an integer, which is the size of the
-     * desired array, and produces an array of the desired size.  This can be
-     * concisely expressed with an array constructor reference:
-     * <pre>{@code
-     *     Person[] men = people.stream()
-     *                          .filter(p -> p.getGender() == MALE)
-     *                          .toArray(Person[]::new);
-     * }</pre>
-     */
-    @Override
-    public <A> A[] toArray(IntFunction<A[]> generator) {
-        return stream.toArray(generator);
-    }
-
-    /**
-     * Performs a <a href="package-summary.html#Reduction">reduction</a> on the
-     * elements of this stream, using the provided identity value and an
-     * <a href="package-summary.html#Associativity">associative</a>
-     * accumulation function, and returns the reduced value.  This is equivalent
-     * to:
-     * <pre>{@code
-     *     T result = identity;
-     *     for (T element : this stream)
-     *         result = accumulator.apply(result, element)
-     *     return result;
-     * }</pre>
-     * <p>
-     * but is not constrained to execute sequentially.
-     *
-     * <p>The {@code identity} value must be an identity for the accumulator
-     * function. This means that for all {@code t},
-     * {@code accumulator.apply(identity, t)} is equal to {@code t}.
-     * The {@code accumulator} function must be an
-     * <a href="package-summary.html#Associativity">associative</a> function.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @param identity    the identity value for the accumulating function
-     * @param accumulator an <a href="package-summary.html#Associativity">associative</a>,
-     *                    <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                    <a href="package-summary.html#Statelessness">stateless</a>
-     *                    function for combining two values
-     * @return the result of the reduction
-     * @apiNote Sum, min, max, average, and string concatenation are all special
-     * cases of reduction. Summing a stream of numbers can be expressed as:
-     *
-     * <pre>{@code
-     *     Integer sum = integers.reduce(0, (a, b) -> a+b);
-     * }</pre>
-     * <p>
-     * or:
-     *
-     * <pre>{@code
-     *     Integer sum = integers.reduce(0, Integer::sum);
-     * }</pre>
-     *
-     * <p>While this may seem a more roundabout way to perform an aggregation
-     * compared to simply mutating a running total in a loop, reduction
-     * operations parallelize more gracefully, without needing additional
-     * synchronization and with greatly reduced risk of data races.
-     */
-    @Override
-    public T reduce(T identity, BinaryOperator<T> accumulator) {
-        return stream.reduce(identity, accumulator);
-    }
-
-    /**
-     * Performs a <a href="package-summary.html#Reduction">reduction</a> on the
-     * elements of this stream, using an
-     * <a href="package-summary.html#Associativity">associative</a> accumulation
-     * function, and returns an {@code Optional} describing the reduced value,
-     * if any. This is equivalent to:
-     * <pre>{@code
-     *     boolean foundAny = false;
-     *     T result = null;
-     *     for (T element : this stream) {
-     *         if (!foundAny) {
-     *             foundAny = true;
-     *             result = element;
-     *         }
-     *         else
-     *             result = accumulator.apply(result, element);
-     *     }
-     *     return foundAny ? Optional.of(result) : Optional.empty();
-     * }</pre>
-     * <p>
-     * but is not constrained to execute sequentially.
-     *
-     * <p>The {@code accumulator} function must be an
-     * <a href="package-summary.html#Associativity">associative</a> function.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @param accumulator an <a href="package-summary.html#Associativity">associative</a>,
-     *                    <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                    <a href="package-summary.html#Statelessness">stateless</a>
-     *                    function for combining two values
-     * @return an {@link Optional} describing the result of the reduction
-     * @throws NullPointerException if the result of the reduction is null
-     * @see #reduce(Object, BinaryOperator)
-     * @see #min(Comparator)
-     * @see #max(Comparator)
-     */
-    @Override
-    public Optional<T> reduce(BinaryOperator<T> accumulator) {
-        return stream.reduce(accumulator);
-    }
-
-    /**
-     * Performs a <a href="package-summary.html#Reduction">reduction</a> on the
-     * elements of this stream, using the provided identity, accumulation and
-     * combining functions.  This is equivalent to:
-     * <pre>{@code
-     *     U result = identity;
-     *     for (T element : this stream)
-     *         result = accumulator.apply(result, element)
-     *     return result;
-     * }</pre>
-     * <p>
-     * but is not constrained to execute sequentially.
-     *
-     * <p>The {@code identity} value must be an identity for the combiner
-     * function.  This means that for all {@code u}, {@code combiner(identity, u)}
-     * is equal to {@code u}.  Additionally, the {@code combiner} function
-     * must be compatible with the {@code accumulator} function; for all
-     * {@code u} and {@code t}, the following must hold:
-     * <pre>{@code
-     *     combiner.apply(u, accumulator.apply(identity, t)) == accumulator.apply(u, t)
-     * }</pre>
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @param identity    the identity value for the combiner function
-     * @param accumulator an <a href="package-summary.html#Associativity">associative</a>,
-     *                    <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                    <a href="package-summary.html#Statelessness">stateless</a>
-     *                    function for incorporating an additional element into a result
-     * @param combiner    an <a href="package-summary.html#Associativity">associative</a>,
-     *                    <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                    <a href="package-summary.html#Statelessness">stateless</a>
-     *                    function for combining two values, which must be
-     *                    compatible with the accumulator function
-     * @return the result of the reduction
-     * @apiNote Many reductions using this form can be represented more simply
-     * by an explicit combination of {@code map} and {@code reduce} operations.
-     * The {@code accumulator} function acts as a fused mapper and accumulator,
-     * which can sometimes be more efficient than separate mapping and reduction,
-     * such as when knowing the previously reduced value allows you to avoid
-     * some computation.
-     * @see #reduce(BinaryOperator)
-     * @see #reduce(Object, BinaryOperator)
-     */
-    @Override
-    public <U> U reduce(U identity, BiFunction<U, ? super T, U> accumulator, BinaryOperator<U> combiner) {
-        return stream.reduce(identity, accumulator, combiner);
-    }
-
-    /**
-     * Performs a <a href="package-summary.html#MutableReduction">mutable
-     * reduction</a> operation on the elements of this stream.  A mutable
-     * reduction is one in which the reduced value is a mutable result container,
-     * such as an {@code ArrayList}, and elements are incorporated by updating
-     * the state of the result rather than by replacing the result.  This
-     * produces a result equivalent to:
-     * <pre>{@code
-     *     R result = supplier.get();
-     *     for (T element : this stream)
-     *         accumulator.accept(result, element);
-     *     return result;
-     * }</pre>
-     *
-     * <p>Like {@link #reduce(Object, BinaryOperator)}, {@code collect} operations
-     * can be parallelized without requiring additional synchronization.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @param supplier    a function that creates a new result container. For a
-     *                    parallel execution, this function may be called
-     *                    multiple times and must return a fresh value each time.
-     * @param accumulator an <a href="package-summary.html#Associativity">associative</a>,
-     *                    <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                    <a href="package-summary.html#Statelessness">stateless</a>
-     *                    function for incorporating an additional element into a result
-     * @param combiner    an <a href="package-summary.html#Associativity">associative</a>,
-     *                    <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                    <a href="package-summary.html#Statelessness">stateless</a>
-     *                    function for combining two values, which must be
-     *                    compatible with the accumulator function
-     * @return the result of the reduction
-     * @apiNote There are many existing classes in the JDK whose signatures are
-     * well-suited for use with method references as arguments to {@code collect()}.
-     * For example, the following will accumulate strings into an {@code ArrayList}:
-     * <pre>{@code
-     *     List<String> asList = stringBaseSteam.collect(ArrayList::new, ArrayList::add,
-     *                                                ArrayList::addAll);
-     * }</pre>
-     *
-     * <p>The following will take a stream of strings and concatenates them into a
-     * single string:
-     * <pre>{@code
-     *     String concat = stringBaseSteam.collect(StringBuilder::new, StringBuilder::append,
-     *                                          StringBuilder::append)
-     *                                 .toString();
-     * }</pre>
-     */
-    @Override
-    public <R> R collect(Supplier<R> supplier, BiConsumer<R, ? super T> accumulator, BiConsumer<R, R> combiner) {
-        return stream.collect(supplier, accumulator, combiner);
-    }
-
-    /**
-     * Performs a <a href="package-summary.html#MutableReduction">mutable
-     * reduction</a> operation on the elements of this stream using a
-     * {@code Collector}.  A {@code Collector}
-     * encapsulates the functions used as arguments to
-     * {@link #collect(Supplier, BiConsumer, BiConsumer)}, allowing for reuse of
-     * collection strategies and composition of collect operations such as
-     * multiple-level grouping or partitioning.
-     *
-     * <p>If the stream is parallel, and the {@code Collector}
-     * is {@link Collector.Characteristics#CONCURRENT concurrent}, and
-     * either the stream is unordered or the collector is
-     * {@link Collector.Characteristics#UNORDERED unordered},
-     * then a concurrent reduction will be performed (see {@link Collector} for
-     * details on concurrent reduction.)
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * <p>When executed in parallel, multiple intermediate results may be
-     * instantiated, populated, and merged so as to maintain isolation of
-     * mutable data structures.  Therefore, even when executed in parallel
-     * with non-thread-safe data structures (such as {@code ArrayList}), no
-     * additional synchronization is needed for a parallel reduction.
-     *
-     * @param collector the {@code Collector} describing the reduction
-     * @return the result of the reduction
-     * @apiNote The following will accumulate strings into an ArrayList:
-     * <pre>{@code
-     *     List<String> asList = stringBaseSteam.collect(Collectors.toList());
-     * }</pre>
-     *
-     * <p>The following will classify {@code Person} objects by city:
-     * <pre>{@code
-     *     Map<String, List<Person>> peopleByCity
-     *         = personBaseSteam.collect(Collectors.groupingBy(Person::getCity));
-     * }</pre>
-     *
-     * <p>The following will classify {@code Person} objects by state and city,
-     * cascading two {@code Collector}s together:
-     * <pre>{@code
-     *     Map<String, Map<String, List<Person>>> peopleByStateAndCity
-     *         = personBaseSteam.collect(Collectors.groupingBy(Person::getState,
-     *                                                      Collectors.groupingBy(Person::getCity)));
-     * }</pre>
-     * @see #collect(Supplier, BiConsumer, BiConsumer)
-     * @see Collectors
-     */
-    @Override
-    public <R, A> R collect(Collector<? super T, A, R> collector) {
-        return stream.collect(collector);
-    }
-
-    /**
-     * Returns the minimum element of this stream according to the provided
-     * {@code Comparator}.  This is a special case of a
-     * <a href="package-summary.html#Reduction">reduction</a>.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal operation</a>.
-     *
-     * @param comparator a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                   <a href="package-summary.html#Statelessness">stateless</a>
-     *                   {@code Comparator} to compare elements of this stream
-     * @return an {@code Optional} describing the minimum element of this stream,
-     * or an empty {@code Optional} if the stream is empty
-     * @throws NullPointerException if the minimum element is null
-     */
-    @Override
-    public Optional<T> min(Comparator<? super T> comparator) {
-        return stream.min(comparator);
-    }
-
-    /**
-     * Returns the maximum element of this stream according to the provided
-     * {@code Comparator}.  This is a special case of a
-     * <a href="package-summary.html#Reduction">reduction</a>.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @param comparator a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                   <a href="package-summary.html#Statelessness">stateless</a>
-     *                   {@code Comparator} to compare elements of this stream
-     * @return an {@code Optional} describing the maximum element of this stream,
-     * or an empty {@code Optional} if the stream is empty
-     * @throws NullPointerException if the maximum element is null
-     */
-    @Override
-    public Optional<T> max(Comparator<? super T> comparator) {
-        return stream.max(comparator);
-    }
-
-    /**
-     * Returns the count of elements in this stream.  This is a special case of
-     * a <a href="package-summary.html#Reduction">reduction</a> and is
-     * equivalent to:
-     * <pre>{@code
-     *     return mapToLong(e -> 1L).sum();
-     * }</pre>
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal operation</a>.
-     *
-     * @return the count of elements in this stream
-     */
-    @Override
-    public long count() {
-        return stream.count();
-    }
-
-    /**
-     * Returns whether any elements of this stream match the provided
-     * predicate.  May not evaluate the predicate on all elements if not
-     * necessary for determining the result.  If the stream is empty then
-     * {@code false} is returned and the predicate is not evaluated.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">short-circuiting
-     * terminal operation</a>.
-     *
-     * @param predicate a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                  <a href="package-summary.html#Statelessness">stateless</a>
-     *                  predicate to apply to elements of this stream
-     * @return {@code true} if any elements of the stream match the provided
-     * predicate, otherwise {@code false}
-     * @apiNote This method evaluates the <em>existential quantification</em> of the
-     * predicate over the elements of the stream (for some x P(x)).
-     */
-    @Override
-    public boolean anyMatch(Predicate<? super T> predicate) {
-        return stream.anyMatch(predicate);
-    }
-
-    /**
-     * Returns whether all elements of this stream match the provided predicate.
-     * May not evaluate the predicate on all elements if not necessary for
-     * determining the result.  If the stream is empty then {@code true} is
-     * returned and the predicate is not evaluated.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">short-circuiting
-     * terminal operation</a>.
-     *
-     * @param predicate a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                  <a href="package-summary.html#Statelessness">stateless</a>
-     *                  predicate to apply to elements of this stream
-     * @return {@code true} if either all elements of the stream match the
-     * provided predicate or the stream is empty, otherwise {@code false}
-     * @apiNote This method evaluates the <em>universal quantification</em> of the
-     * predicate over the elements of the stream (for all x P(x)).  If the
-     * stream is empty, the quantification is said to be <em>vacuously
-     * satisfied</em> and is always {@code true} (regardless of P(x)).
-     */
-    @Override
-    public boolean allMatch(Predicate<? super T> predicate) {
-        return stream.allMatch(predicate);
-    }
-
-    /**
-     * Returns whether no elements of this stream match the provided predicate.
-     * May not evaluate the predicate on all elements if not necessary for
-     * determining the result.  If the stream is empty then {@code true} is
-     * returned and the predicate is not evaluated.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">short-circuiting
-     * terminal operation</a>.
-     *
-     * @param predicate a <a href="package-summary.html#NonInterference">non-interfering</a>,
-     *                  <a href="package-summary.html#Statelessness">stateless</a>
-     *                  predicate to apply to elements of this stream
-     * @return {@code true} if either no elements of the stream match the
-     * provided predicate or the stream is empty, otherwise {@code false}
-     * @apiNote This method evaluates the <em>universal quantification</em> of the
-     * negated predicate over the elements of the stream (for all x ~P(x)).  If
-     * the stream is empty, the quantification is said to be vacuously satisfied
-     * and is always {@code true}, regardless of P(x).
-     */
-    @Override
-    public boolean noneMatch(Predicate<? super T> predicate) {
-        return stream.noneMatch(predicate);
-    }
-
-    /**
-     * Returns an {@link Optional} describing the first element of this stream,
-     * or an empty {@code Optional} if the stream is empty.  If the stream has
-     * no encounter order, then any element may be returned.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">short-circuiting
-     * terminal operation</a>.
-     *
-     * @return an {@code Optional} describing the first element of this stream,
-     * or an empty {@code Optional} if the stream is empty
-     * @throws NullPointerException if the element selected is null
-     */
-    @Override
-    public Optional<T> findFirst() {
-        return stream.findFirst();
-    }
-
-    public Optional<T> findLast() {
-        return Optional.of(toList()).filter(l -> !l.isEmpty()).map(l -> l.get(l.size() - 1));
-    }
-
-    public FastStream<T> reverse() {
-        List<T> list = toList();
-        Collections.reverse(list);
-        return FastStream.of(list);
-    }
-
-    /**
-     * Returns an {@link Optional} describing some element of the stream, or an
-     * empty {@code Optional} if the stream is empty.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">short-circuiting
-     * terminal operation</a>.
-     *
-     * <p>The behavior of this operation is explicitly nondeterministic; it is
-     * free to select any element in the stream.  This is to allow for maximal
-     * performance in parallel operations; the cost is that multiple invocations
-     * on the same source may not return the same result.  (If a stable result
-     * is desired, use {@link #findFirst()} instead.)
-     *
-     * @return an {@code Optional} describing some element of this stream, or an
-     * empty {@code Optional} if the stream is empty
-     * @throws NullPointerException if the element selected is null
-     * @see #findFirst()
-     */
-    @Override
-    public Optional<T> findAny() {
-        return stream.findAny();
-    }
-
-    /**
-     * Returns an iterator for the elements of this stream.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @return the element iterator for this stream
-     */
-    @Override
-    public Iterator<T> iterator() {
-        return stream.iterator();
-    }
-
-    /**
-     * Returns a spliterator for the elements of this stream.
-     *
-     * <p>This is a <a href="package-summary.html#BaseSteamOps">terminal
-     * operation</a>.
-     *
-     * @return the element spliterator for this stream
-     */
-    @Override
-    public Spliterator<T> spliterator() {
-        return stream.spliterator();
-    }
-
-    /**
-     * Returns an equivalent stream that is parallel.  May return
-     * itself, either because the stream was already parallel, or because
-     * the underlying stream state was modified to be parallel.
-     *
-     * <p>This is an <a href="package-summary.html#StreamOps">intermediate
-     * operation</a>.
-     *
-     * @return a parallel stream
-     */
-    @Override
-    public FastStream<T> parallel() {
-        return new FastStream<>(stream.parallel());
-    }
-
-    public FastStream<T> parallel(boolean parallel) {
-        return new FastStream<>(parallel ? stream.parallel() : stream.sequential());
-    }
-
-    // Extra methods for convenience
-
-    /**
-     * Returns an equivalent stream that is
-     * <a href="package-summary.html#Ordering">unordered</a>.  May return
-     * itself, either because the stream was already unordered, or because
-     * the underlying stream state was modified to be unordered.
-     *
-     * <p>This is an <a href="package-summary.html#StreamOps">intermediate
-     * operation</a>.
-     *
-     * @return an unordered stream
-     */
-    @Override
-    public FastStream<T> unordered() {
-        return new FastStream<>(stream.unordered());
-    }
-
-    // Static factories
-
-    /**
-     * Returns an equivalent stream with an additional close handler.  Close
-     * handlers are run when the {@link #close()} method
-     * is called on the stream, and are executed in the order they were
-     * added.  All close handlers are run, even if earlier close handlers throw
-     * exceptions.  If any close handler throws an exception, the first
-     * exception thrown will be relayed to the caller of {@code close()}, with
-     * any remaining exceptions added to that exception as suppressed exceptions
-     * (unless one of the remaining exceptions is the same exception as the
-     * first exception, since an exception cannot suppress itself.)  May
-     * return itself.
-     *
-     * <p>This is an <a href="package-summary.html#StreamOps">intermediate
-     * operation</a>.
-     *
-     * @param closeHandler A task to execute when the stream is closed
-     * @return a stream with a handler that is run if the stream is closed
-     */
-    @Override
-    public FastStream<T> onClose(Runnable closeHandler) {
-        return new FastStream<>(stream.onClose(closeHandler));
-    }
-
-    public FastStream<T> push(T obj) {
-        return FastStream.concat(this, FastStream.of(obj));
-    }
-
-    /**
-     * Returns whether this stream, if a terminal operation were to be executed,
-     * would execute in parallel.  Calling this method after invoking an
-     * terminal stream operation method may yield unpredictable results.
-     *
-     * @return {@code true} if this stream would execute in parallel if executed
-     */
-    @Override
-    public boolean isParallel() {
-        return stream.isParallel();
-    }
-
-    /**
-     * Closes this stream, causing all close handlers for this stream pipeline
-     * to be called.
-     *
-     * @see AutoCloseable#close()
-     */
-    @Override
-    public void close() {
-        stream.close();
-    }
-
-    /**
-     * Returns a hash code value for the object. This method is
-     * supported for the benefit of hash tables such as those provided by
-     * {@link HashMap}.
-     * <p>
-     * The general contract of {@code hashCode} is:
-     * <ul>
-     * <li>Whenever it is invoked on the same object more than once during
-     *     an execution of a Java application, the {@code hashCode} method
-     *     must consistently return the same integer, provided no information
-     *     used in {@code equals} comparisons on the object is modified.
-     *     This integer need not remain consistent from one execution of an
-     *     application to another execution of the same application.
-     * <li>If two objects are equal according to the {@code equals(Object)}
-     *     method, then calling the {@code hashCode} method on each of
-     *     the two objects must produce the same integer result.
-     * <li>It is <em>not</em> required that if two objects are unequal
-     *     according to the {@link Object#equals(Object)}
-     *     method, then calling the {@code hashCode} method on each of the
-     *     two objects must produce distinct integer results.  However, the
-     *     programmer should be aware that producing distinct integer results
-     *     for unequal objects may improve the performance of hash tables.
-     * </ul>
-     * <p>
-     * As much as is reasonably practical, the hashCode method defined by
-     * class {@code Object} does return distinct integers for distinct
-     * objects. (This is typically implemented by converting the internal
-     * address of the object into an integer, but this implementation
-     * technique is not required by the
-     * Java&trade; programming language.)
-     *
-     * @return a hash code value for this object.
-     * @see Object#equals(Object)
-     * @see System#identityHashCode
-     */
-    @Override
-    public int hashCode() {
-        return stream.hashCode();
-    }
-
-    /**
-     * Indicates whether some other object is "equal to" this one.
-     * <p>
-     * The {@code equals} method implements an equivalence relation
-     * on non-null object references:
-     * <ul>
-     * <li>It is <i>reflexive</i>: for any non-null reference value
-     *     {@code x}, {@code x.equals(x)} should return
-     *     {@code true}.
-     * <li>It is <i>symmetric</i>: for any non-null reference values
-     *     {@code x} and {@code y}, {@code x.equals(y)}
-     *     should return {@code true} if and only if
-     *     {@code y.equals(x)} returns {@code true}.
-     * <li>It is <i>transitive</i>: for any non-null reference values
-     *     {@code x}, {@code y}, and {@code z}, if
-     *     {@code x.equals(y)} returns {@code true} and
-     *     {@code y.equals(z)} returns {@code true}, then
-     *     {@code x.equals(z)} should return {@code true}.
-     * <li>It is <i>consistent</i>: for any non-null reference values
-     *     {@code x} and {@code y}, multiple invocations of
-     *     {@code x.equals(y)} consistently return {@code true}
-     *     or consistently return {@code false}, provided no
-     *     information used in {@code equals} comparisons on the
-     *     objects is modified.
-     * <li>For any non-null reference value {@code x},
-     *     {@code x.equals(null)} should return {@code false}.
-     * </ul>
-     * <p>
-     * The {@code equals} method for class {@code Object} implements
-     * the most discriminating possible equivalence relation on objects;
-     * that is, for any non-null reference values {@code x} and
-     * {@code y}, this method returns {@code true} if and only
-     * if {@code x} and {@code y} refer to the same object
-     * ({@code x == y} has the value {@code true}).
-     * <p>
-     * Note that it is generally necessary to override the {@code hashCode}
-     * method whenever this method is overridden, so as to maintain the
-     * general contract for the {@code hashCode} method, which states
-     * that equal objects must have equal hash codes.
-     *
-     * @param obj the reference object with which to compare.
-     * @return {@code true} if this object is the same as the obj
-     * argument; {@code false} otherwise.
-     * @see #hashCode()
-     * @see HashMap
-     */
-    @Override
-    public boolean equals(Object obj) {
-        return stream.equals(obj);
-    }
-
-    /**
-     * Returns a string representation of the object. In general, the
-     * {@code toString} method returns a string that
-     * "textually represents" this object. The result should
-     * be a concise but informative representation that is easy for a
-     * person to read.
-     * It is recommended that all subclasses override this method.
-     * <p>
-     * The {@code toString} method for class {@code Object}
-     * returns a string consisting of the name of the class of which the
-     * object is an instance, the at-sign character `{@code @}', and
-     * the unsigned hexadecimal representation of the hash code of the
-     * object. In other words, this method returns a string equal to the
-     * value of:
-     * <blockquote>
-     * <pre>
-     * getClass().getName() + '@' + Integer.toHexString(hashCode())
-     * </pre></blockquote>
-     *
-     * @return a string representation of the object.
-     */
-    @Override
-    public String toString() {
-        return stream.toString();
-    }
-
-    public <C extends Collection<T>> C toColl(Supplier<C> collectionFactory) {
-        return collect(Collectors.toCollection(collectionFactory));
-    }
-
-    public List<T> toList() {
-        return collect(Collectors.toList());
-    }
-
-    public Set<T> toSet() {
-        return collect(Collectors.toSet());
-    }
-
-    public <R> Map<T, R> toZip(Iterable<R> other) {
-        Iterator<R> iterator = other.iterator();
-        return toMap(Function.identity(), e -> iterator.hasNext() ? iterator.next() : null);
-    }
-
-    public String join() {
-        return join("");
-    }
-
-    public String join(CharSequence delimiter) {
-        return join(delimiter, "", "");
-    }
-
-    public String join(CharSequence delimiter,
-                       CharSequence prefix,
-                       CharSequence suffix) {
-        return map(String::valueOf).collect(Collectors.joining(delimiter, prefix, suffix));
-    }
-
-    public <K> Map<K, T> toMap(Function<? super T, ? extends K> keyMapper) {
-        return toMap(keyMapper, Function.identity());
-    }
-
-    public <K, U> Map<K, U> toMap(Function<? super T, ? extends K> keyMapper,
-                                  Function<? super T, ? extends U> valueMapper) {
-        return toMap(keyMapper, valueMapper, (l, r) -> r);
-    }
-
-    public <K, U> Map<K, U> toMap(Function<? super T, ? extends K> keyMapper,
-                                  Function<? super T, ? extends U> valueMapper,
-                                  BinaryOperator<U> mergeFunction) {
-        return toMap(keyMapper, valueMapper, mergeFunction, HashMap::new);
-    }
-
-    public <K, U, M extends Map<K, U>> M toMap(Function<? super T, ? extends K> keyMapper,
-                                               Function<? super T, ? extends U> valueMapper,
-                                               BinaryOperator<U> mergeFunction,
-                                               Supplier<M> mapSupplier) {
-        return collect(Collectors.toMap(keyMapper, valueMapper, mergeFunction, mapSupplier));
-    }
-
-    public <K> Map<K, List<T>> group(Function<? super T, ? extends K> classifier) {
-        return group(classifier, Collectors.toList());
-    }
-
-    public <K, A, D> Map<K, D> group(Function<? super T, ? extends K> classifier,
-                                     Collector<? super T, A, D> downstream) {
-        return group(classifier, HashMap::new, downstream);
-    }
-
-    public <K, D, A, M extends Map<K, D>> M group(Function<? super T, ? extends K> classifier,
-                                                  Supplier<M> mapFactory,
-                                                  Collector<? super T, A, D> downstream) {
-        return collect(Collectors.groupingBy(classifier, mapFactory, downstream));
-    }
-
-    @SuppressWarnings("unchecked")
-    public FastStream<T> push(T... obj) {
-        return FastStream.concat(this, FastStream.of(obj));
-    }
-
-    public FastStream<T> unshift(T obj) {
-        return FastStream.concat(FastStream.of(obj), this);
-    }
-
-    @SuppressWarnings("unchecked")
-    public FastStream<T> unshift(T... obj) {
-        return FastStream.concat(FastStream.of(obj), this);
-    }
-
-    public T at(Integer idx) {
-        if (Objects.isNull(idx)) {
-            return null;
-        }
-        List<T> list = toList();
-        if (idx > -1) {
-            if (idx >= list.size()) {
-                return null;
-            }
-            return list.get(idx);
-        }
-        if (-idx > list.size()) {
-            return null;
-        }
-        return list.get(list.size() + idx);
-    }
-
-    public T findFirst(Predicate<? super T> predicate) {
-        return filter(predicate).findFirst().orElse(null);
-    }
-
-    public Integer findFirstIdx(Predicate<? super T> predicate) {
-        AtomicInteger idxRef = new AtomicInteger(-1);
-        forEachIdx((e, i) -> {
-            if (predicate.test(e) && idxRef.get() == -1) {
-                idxRef.set(i);
-            }
-        });
-        return idxRef.get();
-    }
-
-    public T findLast(Predicate<? super T> predicate) {
-        return reverse().filter(predicate).findFirst().orElse(null);
-    }
-
-    public Integer findLastIdx(Predicate<? super T> predicate) {
-        AtomicInteger idxRef = new AtomicInteger(-1);
-        forEachIdx((e, i) -> {
-            if (predicate.test(e)) {
-                idxRef.set(i);
-            }
-        });
-        return idxRef.get();
-    }
-
-    public <U, R> FastStream<R> zip(Iterable<U> other,
-                                    BiFunction<? super T, ? super U, ? extends R> zipper) {
-        Iterator<U> iterator = other.iterator();
-        return new FastStream<>(stream.map(e -> zipper.apply(e, iterator.hasNext() ? iterator.next() : null)));
-    }
-
-    @SuppressWarnings("unchecked")
-    public FastStream<T> splice(int start, int deleteCount, T... items) {
-        List<T> list = toList();
-        if (start > -1) {
-            if (start >= list.size()) {
-                return FastStream.concat(FastStream.of(list), FastStream.of(items));
-            }
-            list.removeAll(list.subList(start, start + deleteCount));
-            list.addAll(start, Arrays.asList(items));
-            return FastStream.of(list);
-        }
-        if (-start > list.size()) {
-            return FastStream.concat(FastStream.of(items), FastStream.of(list));
-        }
-        start = list.size() + start;
-        list.removeAll(list.subList(start, start + deleteCount));
-        list.addAll(start, Arrays.asList(items));
-        return FastStream.of(list);
-    }
-
-    public interface StreamBuilder<T> extends Consumer<T> {
-
-        /**
-         * Adds an element to the stream being built.
-         *
-         * @param t the element to add
-         * @throws IllegalStateException if the builder has already transitioned to
-         *                               the built state
-         */
-        @Override
-        void accept(T t);
-
-        /**
-         * Adds an element to the stream being built.
-         *
-         * @param t the element to add
-         * @return {@code this} builder
-         * @throws IllegalStateException if the builder has already transitioned to
-         *                               the built state
-         * @implSpec The default implementation behaves as if:
-         * <pre>{@code
-         *     accept(t)
-         *     return this;
-         * }</pre>
-         */
-        default StreamBuilder<T> add(T t) {
-            accept(t);
-            return this;
-        }
-
-        /**
-         * Builds the stream, transitioning this builder to the built state.
-         * An {@code IllegalStateException} is thrown if there are further attempts
-         * to operate on the builder after it has entered the built state.
-         *
-         * @return the built stream
-         * @throws IllegalStateException if the builder has already transitioned to
-         *                               the built state
-         */
-        FastStream<T> build();
-
-    }
-
-}
